+++ src/catalog_provider/mod.rs
@@ -6,4 +6,6 @@ pub mod glue;
 #[cfg(feature = "deltalake")]
 mod delta_table;
 
+mod glue_schema_provider;
+mod glue_catalog_provider;
 mod glue_table;
